\name{Bayesian}
\alias{read.bay}
\alias{plot.bay}
\alias{pairs.bay}
\alias{print.bay}



\title{Bayesian output analysis}
\description{
  Using Application=BAYES in FEMME a .BAY file is generated which
  can be read and further processed with these functions
}

\usage{
read.bay(bayfile)
\method{plot}{bay}(x, \dots)
\method{pairs}{bay}(x,start,end,\dots)

}

\arguments{
  \item{bayfile}{File ending in .bay generated by FEMME}
  \item{x}{Bayesian object read by \code{read.bay}}
  \item{start}{Skip simulations before start}
  \item{end}{Skip simulations after end}
  \item{...}{Additional plot parameters}
}
\details{

}
\value{
\code{read.bay} returns a list with components
  \item{data}{An object of class coda which can be further processed
    with functions from this package}
  \item{pars}{Parameter names}
  \item{filename}{Filename}
}
\references{Soetaert et al, 2002}
\author{Henrik Andersson <h.andersson@nioo.knaw.nl>, Andreas Hofmann <a.hofmann@nioo.knaw.nl>}
\note{Use at your own risk}


\examples{


## For examples see: vignette("femmeR")

}

\keyword{file}
\keyword{misc}
\keyword{hplot}

